demo/api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.9
opensearch-py==2.7.1
pypdf==5.1.0
python-docx==1.1.2